Add Makefile and get tests working
The Makefile standardizes the install instructions across all of our
Python repos.

Some tests were referencing variables that might not have existed if
a command that ran before their declaration threw an error. Instead
define the variables to the null value and don't run the cleanups if
the variables are null.

Remove a package from update_requirements.txt that cannot be
installed.

Add instructions to the README on how to install dependencies.