get_version_from_setup_py should return string
That's the contract of the function and the other get_version methods
return strings, so we should also return strings.

I didn't notice the earlier test fix because for some reason the test
freenome_build import was loading from the installed version, so
`python setup.py install` was not picking up changes I had made.