Run different install script
Previously we attempted to install from tip when building a Git tag
commit in Travis CI. However, we wouldn't actually install anything,
and the presence of the source files on disk wasn't enough to count as
an install.

Instead run "python setup.py install" if not running the "anaconda
install" process so files get installed to the correct place.

Indent the build script after each pushd so it's clear we're changing
directory, and make it visually clear which working directory
corresponds with which command.